python_projects/ai_diagnostico/requirements.txt
openai==1.12.0
python-dotenv==1.0.1
typing-extensions==4.9.0
